--- src/generate-result/applicators/apply-reroll.ts
@@ -1,49 +1,0 @@
-import { RerollOptions } from 'types'
-
-function rerollRoll(
-  roll: number,
-  { greaterThan, lessThan, exact, maxReroll }: RerollOptions<number>,
-  rollOne: () => number,
-  index = 0
-): number {
-  if (maxReroll === index) {
-    return roll
-  }
-  if (index === 99) {
-    return roll
-  }
-
-  const exactValue =
-    exact !== undefined && Array.isArray(exact)
-      ? exact.includes(roll)
-      : exact === roll
-  if (
-    (greaterThan !== undefined && roll > greaterThan) ||
-    (lessThan !== undefined && roll < lessThan) ||
-    exactValue
-  ) {
-    return rerollRoll(
-      rollOne(),
-      { greaterThan, lessThan, exact, maxReroll },
-      rollOne,
-      index + 1
-    )
-  }
-  return roll
-}
-
-export default function applyReroll(
-  rolls: number[],
-  reroll: RerollOptions<number> | Array<RerollOptions<number>>,
-  rollOne: () => number
-): number[] {
-  const parameters = Array.isArray(reroll) ? reroll : [reroll]
-
-  let rerollRolls = rolls
-  parameters.forEach((rerollModifier) => {
-    rerollRolls = rerollRolls.map((roll) =>
-      rerollRoll(roll, rerollModifier, rollOne)
-    )
-  })
-  return rerollRolls
-}
