--- src/generate-result/applicators/apply-single-cap.ts
@@ -1,16 +1,0 @@
-import { GreaterLessOptions } from 'types'
-
-export default function applySingleCap(
-  { greaterThan, lessThan }: GreaterLessOptions<number>,
-  value?: number
-) {
-  return (roll: number) => {
-    if (greaterThan !== undefined && roll > greaterThan) {
-      return value ?? greaterThan
-    }
-    if (lessThan !== undefined && roll < lessThan) {
-      return value ?? lessThan
-    }
-    return roll
-  }
-}
