--- src/parse-arguments/convertOptionsToParameters/normalizeModifiers/convert-drop-options-to-parameters.ts
@@ -1,17 +1,0 @@
-import { DropOptions } from 'types'
-
-export default function convertDropOptionsToParameters({
-  highest,
-  lowest,
-  greaterThan,
-  lessThan,
-  exact
-}: DropOptions): DropOptions<number> {
-  return {
-    highest: highest === undefined ? undefined : Number(highest),
-    lowest: lowest === undefined ? undefined : Number(lowest),
-    greaterThan: greaterThan === undefined ? undefined : Number(greaterThan),
-    lessThan: lessThan === undefined ? undefined : Number(lessThan),
-    exact: exact === undefined ? undefined : exact.map(Number)
-  }
-}
