--- src/parse-arguments/convertOptionsToParameters/normalizeModifiers/convert-reroll-options-to-parameters.ts
@@ -1,19 +1,0 @@
-import { RerollOptions } from 'types'
-
-import convertGreaterLessOptionsToParameters from './convert-greater-less-options-to-parameters'
-
-export default function convertRerollOptionsToParameters({
-  exact,
-  maxReroll,
-  ...restOptions
-}: RerollOptions): RerollOptions<number> {
-  const convertedExact =
-    exact === undefined
-      ? {}
-      : { exact: Array.isArray(exact) ? exact.map(Number) : [Number(exact)] }
-  return {
-    ...convertGreaterLessOptionsToParameters(restOptions),
-    ...convertedExact,
-    maxReroll: maxReroll === undefined ? undefined : Number(maxReroll)
-  }
-}
