--- src/parse-arguments/parseNotation/notationParsers/parse-cap-notation.ts
@@ -1,22 +1,0 @@
-import { CapMatch, CapModifier } from 'types'
-
-export default function parseCapNotation({
-  capMatch: notationString
-}: CapMatch): CapModifier<number> {
-  let capParameters = {}
-  const capString = notationString.split(/[Cc]/)[1].split(/(?!\d)/)
-  capString.forEach((note) => {
-    if (note.includes('<')) {
-      capParameters = {
-        ...capParameters,
-        lessThan: Number(note.replace(/</g, ''))
-      }
-      return
-    }
-    capParameters = {
-      ...capParameters,
-      greaterThan: Number(note.replace(/>/g, ''))
-    }
-  })
-  return { cap: capParameters }
-}
