--- src/parse-arguments/parseNotation/notationParsers/parse-drop-high-notation.ts
@@ -1,11 +1,0 @@
-import { DropHighMatch, DropModifier } from 'types'
-
-export default function parseDropHighNotation({
-  dropHighMatch: notationString
-}: DropHighMatch): DropModifier<number> {
-  const highestCount = notationString.split(/[Hh]/)[1]
-
-  return {
-    drop: { highest: highestCount === '' ? 1 : Number(highestCount) }
-  }
-}
