--- src/parse-arguments/parseNotation/notationParsers/parse-explode-notation.ts
@@ -1,7 +1,0 @@
-import { ExplodeMatch, ExplodeModifier } from 'types'
-
-export default function parseExplodeNotation({
-  explodeMatch: notationString
-}: ExplodeMatch): ExplodeModifier {
-  return { explode: Boolean(notationString) }
-}
