--- src/parse-arguments/parseNotation/notationParsers/parse-minus-notation.ts
@@ -1,7 +1,0 @@
-import { MinusMatch, MinusModifier } from 'types'
-
-export default function parseMinusNotation({
-  minusMatch: notationString
-}: MinusMatch): MinusModifier<number> {
-  return { minus: Number(notationString.split('-')[1]) }
-}
