--- src/parse-arguments/parseNotation/notationParsers/parse-plus-notation.ts
@@ -1,7 +1,0 @@
-import { PlusMatch, PlusModifier } from 'types'
-
-export default function parsePlusNotation({
-  plusMatch: notationString
-}: PlusMatch): PlusModifier<number> {
-  return { plus: Number(notationString.split('+')[1]) }
-}
